:construction: Get rid of if statements in ingeschreven personen
