+++ src/openpersonen/contrib/stufbg/converters/__init__.py
@@ -1,0 +1,4 @@
+from .ingeschreven_persoon import *  # noqa
+from .kind import *  # noqa
+from .partner import *  # noqa
+from .ouder import *  # noqa
